backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.22
pandas==2.1.4
openai==1.54.0
python-dotenv==1.0.0
httpx==0.27.0
